bot/requirements.txt
# Dependências do Bot de Automação de Notícias

# APIs de IA
anthropic>=0.34.0  # Claude AI
openai>=1.35.0     # DALL-E 3 para imagens

# Processamento de feeds
feedparser>=6.0.11

# Requisições HTTP
requests>=2.31.0

# Utilitários
python-dotenv>=1.0.0  # Para carregar .env

